requirements.txt
g2pk
torch
numpy
tgt
scipy
pyworld
librosa
numba
matplotlib
unidecode
inflect
g2p-en
tensorboard
python_speech_features
pandas
pysptk
tensorflow
PyYAML
tqdm
moviepy
quickspacer